requirements.txt
# Inference — use tflite-runtime on Linux/Docker, tensorflow on Windows
# On Linux (Docker): pip install tflite-runtime==2.14.0
# On Windows (dev):  pip install tensorflow==2.13.0
tensorflow==2.13.0   # covers both inference (tflite) and Grad-CAM on Windows
# For Docker production swap to: tflite-runtime==2.14.0 + tensorflow-cpu==2.13.0
numpy==1.24.3
Pillow==10.0.0

# Grad-CAM image processing (headless = no display server needed in Docker)
opencv-python-headless==4.8.1.78

# Protobuf pinned to avoid firebase-admin MessageFactory AttributeError
protobuf==3.20.3

# API
fastapi==0.103.0
pydantic==1.10.13
uvicorn[standard]==0.23.2
python-multipart==0.0.6

# Rate limiting
slowapi==0.1.9

# History stored in SQLite — no external service needed
# firebase-admin removed (replaced with SQLite)

# HTTP (shops service)
httpx==0.25.0

# Logging
structlog==23.1.0

# Env
python-dotenv==1.0.0
